SNOW-902611 teardown of pool settings changes in tests
