+++ Snowflake.Data.Tests/App.config
@@ -21,7 +21,7 @@ Copyright (c) 2012-2017 Snowflake Computing Inc. All rights reserved.
       </layout>
       <appender name="ConsoleAppender" type="log4net.Appender.ConsoleAppender">
         <layout type="log4net.Layout.PatternLayout">
-          <conversionPattern value="%date [%thread] [%-5level] [%ClassName:%line] - %message%newline" />
+          <conversionPattern value="%date [%thread] [%-5level] [%classname:%line] - %message%newline" />
         </layout>
       </appender>
     </appender>
@@ -31,7 +31,7 @@ Copyright (c) 2012-2017 Snowflake Computing Inc. All rights reserved.
       <appender-ref ref="RollingFileAppender" />
     </root>    
     <root>
-      <level value="WARN" />
+      <level value="ALL" />
       <appender-ref ref="ConsoleAppender" />
     </root>
   </log4net>
